+++ public/views/home/association.php
@@ -9,6 +9,6 @@
             <form method="post"
                 action="#">
             <td class="space"><p><input type="hidden" name="association[name]" value="<?=$association->nickname?>" /></p>
-            <td class="actions"><p><button name="association[redirect]" value="page">Visit page</button><?php if (!$iterator->getSawUsers()): ?><br />
-            <button name="association[redirect]" value="admin">Admnistrator Panel</button><?php endif ?></p>
+            <td class="actions"><p><a href="<?=HOME_URI?>/@<?=$association->nickname?>">Visit page</a><?php if ($association->checkIfAdmin(clone UserSession::getUser())): ?><br />
+            <a href="<?=HOME_URI?>/@<?=$association->nickname?>/admni">Admnistrator Panel</a><?php endif ?></p>
             </form>