+++ public/views/includes/nav.php
@@ -20,6 +20,7 @@ if ($this->loginRequired && !UserSession::getUser()->loggedIn)
         <li><a href="<?=HOME_URI?>/userRegister">Sign In</a></li>
         <li><a href="<?=HOME_URI?>/associations">Associations</a></li>
         <li><a href="<?=HOME_URI?>/news">News</a></li>
+        <li><a href="<?=HOME_URI?>/events">Events</a></li>
         <li><a href="<?=HOME_URI?>/projects">Projetos</a></li>
         <li><a href="<?=HOME_URI?>/projects/adm">Gerir Projetos</a></li>
     </ul>